run_steric_resolution.py now preserves input box dimensions using floating point data rather than integer.

+++ docker_build/run_steric_resolution.py
@@ -244,7 +244,7 @@ def run_steric_resolution_loop(input_coord_file, index_list, residue_names_list,
         adjusted_coords = forcefield_parent_filepath + 'adjusted_coords.gro'
         output_universe.atoms.write(adjusted_coords)
         #adjust the box vectors (which get set to 0 for whatever reason)
-        subprocess.call(['/bin/bash','-i','-c','gmx editconf -f {coord_path} -o {coord_path} -box {dimensions_0} {dimensions_1} {dimensions_2} -noc'.format(coord_path=adjusted_coords, dimensions_0 = int(dimensions[0] / 10.), dimensions_1 = int(dimensions[1] /10.), dimensions_2 = int(dimensions[2] / 10.))])
+        subprocess.call(['/bin/bash','-i','-c','gmx editconf -f {coord_path} -o {coord_path} -box {dimensions_0} {dimensions_1} {dimensions_2} -noc'.format(coord_path=adjusted_coords, dimensions_0 = float(dimensions[0] / 10.), dimensions_1 = float(dimensions[1] /10.), dimensions_2 = float(dimensions[2] / 10.))])
         input_coord_file = adjusted_coords #use the new coord file as the algorithm input
 
         # up next, need to deal with writing new .top and .itp files in preparation for the selective application of position restraints
